+++ src/main/java/com/lgmpjt/websnsspringboot/application/port/service/UserService.java
@@ -1,4 +1,4 @@
-package com.lgmpjt.websnsspringboot.domain.user.service;
+package com.lgmpjt.websnsspringboot.application.port.service;
 
 import com.lgmpjt.websnsspringboot.application.port.in.dto.UserDto;
 import com.lgmpjt.websnsspringboot.adapter.out.entity.User;
