+++ src/test/java/com/lgmpjt/websnsspringboot/follow/FollowApiTest.java
@@ -1,9 +1,9 @@
 package com.lgmpjt.websnsspringboot.follow;
 
 import com.lgmpjt.websnsspringboot.ApiTest;
-import com.lgmpjt.websnsspringboot.domain.follow.service.FollowService;
+import com.lgmpjt.websnsspringboot.application.port.service.FollowService;
 import com.lgmpjt.websnsspringboot.application.port.in.dto.UserCreateDto;
-import com.lgmpjt.websnsspringboot.domain.user.service.UserService;
+import com.lgmpjt.websnsspringboot.application.port.service.UserService;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
